// apps/mobile/src/api/generated/models/diaryQuantityUnit.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Minha Saúde API
 * API do app de nutrição e acompanhamento de saúde.
 * OpenAPI spec version: 0.1.0
 */

export type DiaryQuantityUnit = typeof DiaryQuantityUnit[keyof typeof DiaryQuantityUnit];


export const DiaryQuantityUnit = {
  grams: 'grams',
  portion: 'portion',
} as const;